Encrypt to Filename
So, I was spitting stuff out on stdout when they came in from stdin.
Turns out that was dumb, as I still need to wrap other layers on top, so
it needs to go in a file.

Then, also, I spit it out, so other stuff can get there, once I get
there.