Do The Thing
Ok, so I've got the non-basecase working now.
It actually tars up the files of the lower level, and encrypts them to
the key we're on.
Then it outputs those all to the next level up.

At the top level we extract all the items from the tempdir, put them in
the current folder for sending, etc, and then cleanup.

Roughly, it works.